+++ src/aim/cftemplates/s3.py
@@ -10,18 +10,22 @@ class S3(CFTemplate):
     def __init__(self,
                 aim_ctx,
                 account_ctx,
-                buckets,
-                s3_context_id,
+                bucket_context,
+                bucket_policy_only,
                 config_ref):
+
+        aws_name = '-'.join(['S3', bucket_context['group_id'], bucket_context['id']])
+        if bucket_policy_only == True:
+            aws_name += '-policy'
 
         super().__init__(aim_ctx,
                          account_ctx,
                          config_ref=config_ref,
-                         aws_name='S3',
+                         aws_name=aws_name,
                          iam_capabilities=["CAPABILITY_NAMED_IAM"])
 
-        self.s3_context_id = s3_context_id
-        self.buckets = buckets
+        self.s3_context_id = config_ref
+        self.bucket_context = bucket_context
 
         # Define the Template
         template_fmt = """
@@ -34,7 +38,6 @@ Description: 'S3 Buckets'
 Resources:
 {0[resources_yaml]:s}
 
-Outputs:
 {0[outputs_yaml]:s}
 """
         template_table = {
@@ -53,18 +56,21 @@ Outputs:
 
         s3_policy_fmt = """
   {0[cf_resource_name_prefix]:s}BucketPolicy:
-    DependsOn:
-      - {0[cf_resource_name_prefix]:s}Bucket
+  {0[cf_policy_depends_on]:s}
     Type: AWS::S3::BucketPolicy
     Properties:
       Bucket: {0[bucket_name]:s}
       PolicyDocument:
         Statement: {0[policy_statements]:s}
 """
+        s3_policy_depends_on_fmt = """  DependsOn:
+      - {0[cf_resource_name_prefix]:s}Bucket"""
+
         s3_policy_table = {
             'cf_resource_name_prefix': "",
             'bucket_name' : "",
-            'policy_statements': ""
+            'policy_statements': "",
+            'cf_policy_depends_on': ""
         }
 
         s3_policy_statement_fmt = """
@@ -82,6 +88,7 @@ Outputs:
             }
 
         s3_bucket_outputs_fmt = """
+Outputs:
   {0[cf_resource_name_prefix]:s}BucketName:
     Value: !Ref {0[cf_resource_name_prefix]:s}Bucket
 """
@@ -102,47 +109,56 @@ Outputs:
         outputs_yaml = ""
 
         s3_ctl = self.aim_ctx.get_controller('S3')
-        for bucket_context in self.buckets:
-            s3_bucket_table.clear()
-            bucket_config = bucket_context['config']
-            s3_bucket_table['bucket_name'] = s3_ctl.get_bucket_name(self.s3_context_id, bucket_context['ref'])
-            s3_bucket_table['cf_resource_name_prefix'] = self.gen_cf_logical_name(bucket_context['id'], '_')
 
+        s3_bucket_table.clear()
+        bucket_config = bucket_context['config']
+        s3_bucket_table['bucket_name'] = s3_ctl.get_bucket_name(self.s3_context_id)
+        s3_bucket_table['cf_resource_name_prefix'] = self.gen_cf_logical_name(self.bucket_context['id'], '_')
+
+        if bucket_policy_only == False:
             # parameters_yaml += s3_bucket_params_fmt.format(s3_bucket_table)
             resources_yaml += s3_bucket_fmt.format(s3_bucket_table)
 
-            # Bucket Policy
-            if len(bucket_config.policy) > 0:
-                s3_policy_table['cf_resource_name_prefix'] = s3_bucket_table['cf_resource_name_prefix']
-                s3_policy_table['bucket_name'] = s3_bucket_table['bucket_name']
-                s3_policy_table['policy_statements'] = ""
-                # Statement
-                for policy_statement in bucket_config.policy:
-                    s3_policy_statement_table['action_list'] = ""
-                    s3_policy_statement_table['principal'] = ""
-                    s3_policy_statement_table['effect'] = ""
-                    s3_policy_statement_table['resource_list'] = ""
-                    # Action
-                    for action in policy_statement.action:
-                        s3_policy_statement_table['action_list'] += cf_list_item.format(action)
-                    # Effict
-                    s3_policy_statement_table['effect'] = policy_statement.effect
-                    # Principal
-                    if policy_statement.aws != None and len(policy_statement.aws) > 0:
-                        s3_policy_statement_table['principal'] = "              AWS:"
-                        for principal in policy_statement.aws:
-                            s3_policy_statement_table['principal'] += cf_principal_list_item.format(principal)
+        # Bucket Policy
+        if len(bucket_config.policy) > 0:
+            if not bucket_policy_only:
+                s3_policy_table['cf_policy_depends_on'] = s3_policy_depends_on_fmt.format(s3_bucket_table)
 
-                    # Resource
-                    bucket_arn = s3_ctl.get_bucket_arn(self.s3_context_id, bucket_context['ref'])
-                    if policy_statement.resource_suffix and len(policy_statement.resource_suffix) > 0:
-                        for res_suffix in policy_statement.resource_suffix:
-                            resource_arn = bucket_arn + res_suffix
-                            s3_policy_statement_table['resource_list'] += cf_list_item.format(resource_arn)
-                    else:
-                        s3_policy_statement_table['resource_list'] = cf_list_item.format(bucket_arn)
-                    s3_policy_table['policy_statements'] += s3_policy_statement_fmt.format(s3_policy_statement_table)
-                resources_yaml += s3_policy_fmt.format(s3_policy_table)
+            s3_policy_table['cf_resource_name_prefix'] = s3_bucket_table['cf_resource_name_prefix']
+            s3_policy_table['bucket_name'] = s3_bucket_table['bucket_name']
+            s3_policy_table['policy_statements'] = ""
+            # Statement
+            for policy_statement in bucket_config.policy:
+                if policy_statement.written_to_template == True:
+                    continue
+                else:
+                    policy_statement.written_to_template = True
+                s3_policy_statement_table['action_list'] = ""
+                s3_policy_statement_table['principal'] = ""
+                s3_policy_statement_table['effect'] = ""
+                s3_policy_statement_table['resource_list'] = ""
+                # Action
+                for action in policy_statement.action:
+                    s3_policy_statement_table['action_list'] += cf_list_item.format(action)
+                # Effict
+                s3_policy_statement_table['effect'] = policy_statement.effect
+                # Principal
+                if policy_statement.aws != None and len(policy_statement.aws) > 0:
+                    s3_policy_statement_table['principal'] = "              AWS:"
+                    for principal in policy_statement.aws:
+                        s3_policy_statement_table['principal'] += cf_principal_list_item.format(principal)
+
+                # Resource
+                bucket_arn = s3_ctl.get_bucket_arn(self.s3_context_id)
+                if policy_statement.resource_suffix and len(policy_statement.resource_suffix) > 0:
+                    for res_suffix in policy_statement.resource_suffix:
+                        resource_arn = bucket_arn + res_suffix
+                        s3_policy_statement_table['resource_list'] += cf_list_item.format(resource_arn)
+                else:
+                    s3_policy_statement_table['resource_list'] = cf_list_item.format(bucket_arn)
+                s3_policy_table['policy_statements'] += s3_policy_statement_fmt.format(s3_policy_statement_table)
+            resources_yaml += s3_policy_fmt.format(s3_policy_table)
+        if bucket_policy_only == False:
             outputs_yaml += s3_bucket_outputs_fmt.format(s3_bucket_table)
 
         template_table['parameters_yaml'] = parameters_yaml
@@ -167,6 +183,5 @@ Outputs:
 
     def delete(self):
         s3_ctl = self.aim_ctx.get_controller('S3')
-        for bucket_context in self.buckets:
-            s3_ctl.empty_bucket(self.s3_context_id, bucket_context['ref'])
+        s3_ctl.empty_bucket(self.s3_context_id, self.bucket_context['ref'])
         super().delete()